Fix some warnings from new inspection tool

+++ core/src/main/java/com/google/zxing/datamatrix/encoder/HighLevelEncoder.java
@@ -251,12 +251,12 @@ public final class HighLevelEncoder {
 
       //step L
       if (isDigit(c)) {
-        charCounts[ASCII_ENCODATION] += 0.5;
+        charCounts[ASCII_ENCODATION] += 0.5f;
       } else if (isExtendedASCII(c)) {
-        charCounts[ASCII_ENCODATION] = (int) Math.ceil(charCounts[ASCII_ENCODATION]);
-        charCounts[ASCII_ENCODATION] += 2;
+        charCounts[ASCII_ENCODATION] = (float) Math.ceil(charCounts[ASCII_ENCODATION]);
+        charCounts[ASCII_ENCODATION] += 2.0f;
       } else {
-        charCounts[ASCII_ENCODATION] = (int) Math.ceil(charCounts[ASCII_ENCODATION]);
+        charCounts[ASCII_ENCODATION] = (float) Math.ceil(charCounts[ASCII_ENCODATION]);
         charCounts[ASCII_ENCODATION]++;
       }
 
@@ -298,7 +298,7 @@ public final class HighLevelEncoder {
 
       // step Q
       if (isSpecialB256(c)) {
-        charCounts[BASE256_ENCODATION] += 4;
+        charCounts[BASE256_ENCODATION] += 4.0f;
       } else {
         charCounts[BASE256_ENCODATION]++;
       }
